requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
langchain
langchain-openai
openai
langchain_community
dotenv
google-genai==1.7.0
google-generativeai


